pyproject.toml
[project]
name = "nim-llm-game"
version = "0.1.0"
description = "A Nim game implementation with AI opponent using Large Language Models"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "python-dotenv",
    "ipykernel",
    "ipywidgets",
    "requests",
    "numpy",
    "pandas",
    "scipy",
    "scikit-learn",
    "openai",
    "anthropic",
    "ollama",
    "accelerate",
    "bitsandbytes",
    "gradio",
    "google-genai>=1.41.0",
    "google-generativeai>=0.8.5",
    "groq>=0.33.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]


